+++ opti_oignon/rag/__init__.py
@@ -30,7 +30,7 @@ CLI:
 Author: Léon
 """
 
-__version__ = "2.0.1"
+__version__ = "2.0.2"
 __author__ = "Léon"
 
 # Main imports